create test_utils_plots.py
